accept project name

+++ flask_setup/__init__.py
@@ -20,10 +20,28 @@ def init():
 
 @app.command()
 @before_command
-def build():
+def build(project: str = ''):
+
+    while project == '':
+        project = input('Enter project name: ')
+
+    if os.path.exists(project):
+        typer.echo("Project already exists")
+        return
+    
+    # Create project folder
+    os.mkdir(project)
+    do_add_log("Project folder created")
+    # move to project folder
+    os.chdir(project)
+
     path = os.path.dirname(os.path.realpath(__file__))
     # copy relevant app files to project directory
     copytree(f'{path}/starter', '.', dirs_exist_ok=True)
+    
+    # create virtual environment
+    os.system('python3 -m venv venv') if os.name == 'posix' else os.system('py -m venv venv')
+
     # install requirements
     os.system(f'pip install -r requirements.txt')
     log = 'Project built successfully'
